# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SAAS API",
        "description": "API documentation for SAAS Dashboard",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/create_stage": {
            "post": {
                "tags": [
                    "Stages"
                ],
                "summary": "إنشاء مرحلة جديدة",
                "operationId": "5506135e7544376e11e79ef7f18bcfa7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "education_type_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "المرحلة الإعدادية"
                                    },
                                    "education_type_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "parent_id": {
                                        "type": "integer",
                                        "example": null
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "نجاح العملية",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "تم إنشاء المرحلة بنجاح"
                                        },
                                        "stage": {
                                            "$ref": "#/components/schemas/StageResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "خطأ في التحقق من البيانات",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "حدث خطأ، البيانات غير صالحة"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/update_stage": {
            "post": {
                "tags": [
                    "Stages"
                ],
                "summary": "تحديث مرحلة",
                "operationId": "795d30f1c92a03c9bfe44de18276dc10",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID الخاص بالمرحلة",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "المرحلة الثانوية"
                                    },
                                    "education_type_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "parent_id": {
                                        "type": "integer",
                                        "example": null
                                    },
                                    "subject_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            1,
                                            2
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "نجاح العملية",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "تم تحديث المرحلة بنجاح"
                                        },
                                        "stage": {
                                            "$ref": "#/components/schemas/StageResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "خطأ في التحقق من البيانات",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "حدث خطأ، البيانات غير صالحة"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/fetch_stage": {
            "post": {
                "tags": [
                    "Stages"
                ],
                "summary": "جلب مرحلة معينة",
                "operationId": "4046446117972b9068dca2c7e03baeaf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "نجاح العملية",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "تم جلب المرحلة بنجاح"
                                        },
                                        "stage": {
                                            "$ref": "#/components/schemas/StageResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "المرحلة غير موجودة",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "المرحلة غير موجودة"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/delete_stage": {
            "post": {
                "tags": [
                    "Stages"
                ],
                "summary": "حذف مرحلة",
                "operationId": "085c7b5950abb5806eb2e3ddddb4c52e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "نجاح العملية",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "تم حذف المرحلة بنجاح"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "المرحلة غير موجودة",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "المرحلة غير موجودة"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StageResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "المرحلة الإعدادية"
                    },
                    "education_type_id": {
                        "type": "integer",
                        "example": 3
                    },
                    "organization_id": {
                        "type": "integer",
                        "example": 5
                    },
                    "parent_id": {
                        "type": "integer",
                        "example": 0,
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Use bearer token to access this API",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}